fix: inline the checkpointing helpers — hf jobs uploads ONE script
`hf jobs uv run` takes a single SCRIPT (local file or URL) with no way to ship a
sibling module, so `import checkpointing` would have raised ImportError inside
the container the moment the job started. Caught before launching, not after.

The helpers now live in train.py and the tests load them from there via
importlib rather than keeping a second copy — the duplicate-source trap that
already cost three robot sessions on the r6d convention.

Verified the property that actually matters: train.py copied ALONE into an empty
directory imports and resolves its push targets correctly
(<id>_step4000 / <id>_best). 325 tests pass.
